requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
aiosqlite>=0.19.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dateutil>=2.8.2
openpyxl>=3.1.2
xlrd>=2.0.0
reportlab>=4.0.9
qrcode[pil]>=7.4.2
